api/AI_api/requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.11.0
einops==0.2.0
filelock==3.12.4
fonttools==4.42.1
gdown==4.4.0
gitdb==4.0.10
GitPython==3.1.36
idna==3.4
imageio==2.31.3
imgaug==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
prefetch-generator==1.0.1
psutil==5.9.5
py-cpuinfo==9.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.2
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12
tifffile==2023.9.18
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.181
urllib3==2.0.4
vietocr==0.3.11
Werkzeug==2.3.7
